+++ packages/core/src/server/browserLogs.ts
@@ -2,7 +2,8 @@ import path from 'node:path';
 import { promisify } from 'node:util';
 import { parse as parseStack, type StackFrame } from 'stacktrace-parser';
 import { SCRIPT_REGEX } from '../constants';
-import { color, requireCompiledPackage } from '../helpers';
+import { color } from '../helpers';
+import { requireCompiledPackage } from '../helpers/vendors';
 import { logger } from '../logger';
 import type { EnvironmentContext, InternalContext, Rspack } from '../types';
 import { getFileFromUrl } from './assets-middleware/getFileFromUrl';
